sandbox/dan2: committing a bunch of changes to neural-network training setup which are not yet ready to merge to trunk.
git-svn-id: https://svn.code.sf.net/p/kaldi/code/sandbox/dan2@2777 5e6a8d80-dfce-4ca6-a32a-6e07a63d50c8

+++ egs/wsj/s5/steps/train_lda_mllt.sh
@@ -23,6 +23,7 @@ splice_opts=
 cluster_thresh=-1  # for build-tree control final bottom-up clustering of leaves
 # End configuration.
 train_tree=true  # if false, don't actually train the tree.
+use_lda_mat=  # If supplied, use this LDA[+MLLT] matrix.
 
 echo "$0 $@"  # Print the command line for logging
 
@@ -73,15 +74,24 @@ feats="$splicedfeats transform-feats $dir/0.mat ark:- ark:- |"
 
 
 if [ $stage -le -5 ]; then
-  echo "Accumulating LDA statistics."
-  $cmd JOB=1:$nj $dir/log/lda_acc.JOB.log \
+  if [ -z "$use_lda_mat" ]; then
+    echo "Accumulating LDA statistics."
+    $cmd JOB=1:$nj $dir/log/lda_acc.JOB.log \
     ali-to-post "ark:gunzip -c $alidir/ali.JOB.gz|" ark:- \| \
       weight-silence-post 0.0 $silphonelist $alidir/final.mdl ark:- ark:- \| \
       acc-lda --rand-prune=$randprune $alidir/final.mdl "$splicedfeats" ark,s,cs:- \
-       $dir/lda.JOB.acc || exit 1;
-  est-lda --write-full-matrix=$dir/full.mat --dim=$dim $dir/0.mat $dir/lda.*.acc \
+      $dir/lda.JOB.acc || exit 1;
+    est-lda --write-full-matrix=$dir/full.mat --dim=$dim $dir/0.mat $dir/lda.*.acc \
       2>$dir/log/lda_est.log || exit 1;
-  rm $dir/lda.*.acc
+    rm $dir/lda.*.acc
+  else
+    echo "Using supplied LDA matrix $use_lda_mat"
+    cp $use_lda_mat $dir/0.mat || exit 1;
+    [ ! -z "$mllt_iters" ] && \
+      echo "Warning: using supplied LDA matrix $use_lda_mat but we will do MLLT," && \
+      echo "which you might not want; to disable MLLT, specify --mllt-iters ''" && \
+      sleep 5
+  fi
 fi
 
 cur_lda_iter=0
